Move common_defs.h to the top-level directory
